% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDendro.R
\name{shinyDendro}
\alias{shinyDendro}
\title{shinyDendro}
\usage{
shinyDendro(inputId, cl_height, merge, order, heatmap = NULL,
  assignment = NULL, fontScale = 0.66, fontFg = "#ffffffff",
  fontShadow = "#000000ff", key = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{inputId}{Name of the associated reactive input}

\item{cl_height, merge, order}{Heights, merge matrix and ordering of a dendrogram (as taken e.g. from hclust result).}

\item{heatmap}{Matrix with colors and named columns to display next to the dendrogram (may be updated dynamically)}

\item{assignment}{Initial assignment of the clusters (character vector, may be updated dynamically)}

\item{width, height, elementId}{Widget parameters}
}
\description{
Create an interactive dendrogram.
}
\examples{
\donttest{
cl <- hclust(dist(cbind(rnorm(100),rnorm(100))))
renderShinyDendro(shinyDendro('sdClusters', cl$height, cl$merge, cl$order))
}
}
